Issue GH-47: Binary keys for JRedis and JReidsFuture.  Tested.
Review changes in test suite to pin point the API methods that are no
longer backwards compatible.  These are a tiny subset of commands that
would return a key (not a value), such as RANDOMKEY.  Instead of
returning String, they now return byte[].  All other API changes are
backwards compatible so this should be of minimal pain to existing
applications.