Phase 1 of decoupling socket and queue in critical section - TESTED
This is a fairly substantial change to the pipeline queueRequest()
method, introducing a temporary connection queue to reduce interference
between clients and the response handler thread, and, to prepare for
stripping the critical zone into finer sections to reduce overall
contention.

Performance is somewhat (measurable but marginal) better than the
previous versionof the chunked pipeline.  Jitter seems to be lower
but that needs to be actually measured.